# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend da POC do centralizador de saldo. Toda regra temporal usa POC_REFERENCE_DATE.",
        "title": "Centralizador de Saldo (POC) — API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/goals/{goalID}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goals"
                ],
                "summary": "Detalhe de um objetivo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do objetivo",
                        "name": "goalID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.GoalDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/goals/{goalID}/allocations": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goals"
                ],
                "summary": "Adiciona uma alocação (conta-fonte) a um objetivo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do objetivo",
                        "name": "goalID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Conta e percentual",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.AddAllocationRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/goals/{goalID}/allocations/{accountID}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goals"
                ],
                "summary": "Remove uma alocação (conta-fonte) de um objetivo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do objetivo",
                        "name": "goalID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID da conta-fonte",
                        "name": "accountID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/goals/{goalID}/backtest": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backtest"
                ],
                "summary": "Consulta o resultado do backtest",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do objetivo",
                        "name": "goalID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.BacktestResultDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backtest"
                ],
                "summary": "Executa o backtest de um objetivo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do objetivo",
                        "name": "goalID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.BacktestResultDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Lista clientes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtro por substring do nome",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_httpapi.UserDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/{userID}/accounts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Lista contas de um cliente com saldo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do cliente",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_httpapi.AccountDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/{userID}/assistant/plan": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assistant"
                ],
                "summary": "Turno do assistente de planejamento (chat)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do cliente",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Histórico da conversa",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.AssistantPlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.AssistantPlanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/{userID}/goals": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goals"
                ],
                "summary": "Lista objetivos de um cliente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do cliente",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_httpapi.GoalSummaryDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "goals"
                ],
                "summary": "Cria um objetivo (goal + vault + allocations)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do cliente",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do objetivo",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.CreateGoalRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_httpapi.CreateGoalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_httpapi.AccountDTO": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "balance": {
                    "type": "string",
                    "example": "1234.56"
                },
                "balance_reference_date": {
                    "type": "string",
                    "example": "2025-01-01"
                },
                "branch_code": {
                    "type": "string"
                },
                "brand_name": {
                    "type": "string"
                },
                "check_digit": {
                    "type": "string"
                },
                "compe_code": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.AddAllocationRequest": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "percentage": {
                    "type": "integer",
                    "example": 50
                }
            }
        },
        "internal_httpapi.AllocationRequest": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "percentage": {
                    "type": "integer",
                    "example": 50
                }
            }
        },
        "internal_httpapi.AssistantPlanRequest": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_httpapi.ChatMessageDTO"
                    }
                }
            }
        },
        "internal_httpapi.AssistantPlanResponse": {
            "type": "object",
            "properties": {
                "done": {
                    "type": "boolean"
                },
                "proposal": {
                    "$ref": "#/definitions/internal_httpapi.PlanProposalDTO"
                },
                "reply": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.BacktestMovementDTO": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "amount": {
                    "type": "string",
                    "example": "500.00"
                },
                "entradas": {
                    "type": "string",
                    "example": "2000.00"
                },
                "reference_month": {
                    "type": "string",
                    "example": "2025-06-01"
                },
                "saidas": {
                    "type": "string",
                    "example": "1500.00"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.BacktestResultDTO": {
            "type": "object",
            "properties": {
                "movements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_httpapi.BacktestMovementDTO"
                    }
                },
                "summary": {
                    "$ref": "#/definitions/internal_httpapi.BacktestSummaryDTO"
                },
                "vault_evolution": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_httpapi.VaultEvolutionDTO"
                    }
                }
            }
        },
        "internal_httpapi.BacktestSummaryDTO": {
            "type": "object",
            "properties": {
                "completed_count": {
                    "type": "integer"
                },
                "completed_months_pct": {
                    "type": "number",
                    "example": 0.83
                },
                "failed_count": {
                    "type": "integer"
                },
                "goal_reached": {
                    "type": "boolean"
                },
                "target_amount": {
                    "type": "string",
                    "example": "10000.00"
                },
                "vault_balance": {
                    "type": "string",
                    "example": "8400.00"
                },
                "worst_account_id": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.ChatMessageDTO": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Quero juntar 10 mil para uma viagem"
                },
                "role": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
...
        "internal_httpapi.CreateGoalRequest": {
            "type": "object",
            "properties": {
                "allocations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_httpapi.AllocationRequest"
                    }
                },
                "duration_months": {
                    "type": "integer",
                    "example": 12
                },
                "name": {
                    "type": "string",
                    "example": "Viagem"
                },
                "start_date": {
                    "type": "string",
                    "example": "2025-01-01"
                },
                "target_amount": {
                    "type": "string",
                    "example": "10000.00"
                }
            }
        },
        "internal_httpapi.CreateGoalResponse": {
            "type": "object",
            "properties": {
                "goal_id": {
                    "type": "string"
                },
                "vault_id": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.GoalAllocationDTO": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "brand_name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "percentage": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.GoalDTO": {
            "type": "object",
            "properties": {
                "allocations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_httpapi.GoalAllocationDTO"
                    }
                },
                "duration_months": {
                    "type": "integer"
                },
                "goal_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "target_amount": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "vault_id": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.GoalSummaryDTO": {
            "type": "object",
            "properties": {
                "duration_months": {
                    "type": "integer"
                },
                "goal_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "target_amount": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.PlanOptionDTO": {
            "type": "object",
            "properties": {
                "duration_months": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "target_amount": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.PlanProposalDTO": {
            "type": "object",
            "properties": {
                "allocations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_httpapi.ProposedAllocationDTO"
                    }
                },
                "name": {
                    "type": "string"
                },
                "plans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_httpapi.PlanOptionDTO"
                    }
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.ProposedAllocationDTO": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "string"
                },
                "brand_name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "percentage": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.UserDTO": {
            "type": "object",
            "properties": {
                "accounts_count": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_httpapi.VaultEvolutionDTO": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "string",
                    "example": "1000.00"
                },
                "month": {
                    "type": "string",
                    "example": "2025-06-01"
                }
            }
        }
    }
}